[FEATURE] Move shared SQL code to abstract class
Also ports permissions support to MySQL, because it doen`t have PostgreSQL-specific code.
