+++ LaunchServer/src/main/java/pro/gravit/launchserver/auth/PostgreSQLSourceConfig.java
@@ -10,7 +10,7 @@ import javax.sql.DataSource;
 import java.sql.Connection;
 import java.sql.SQLException;
 
-public final class PostgreSQLSourceConfig implements AutoCloseable {
+public final class PostgreSQLSourceConfig implements AutoCloseable, SQLSourceConfig {
     public static final int TIMEOUT = VerifyHelper.verifyInt(
             Integer.parseUnsignedInt(System.getProperty("launcher.postgresql.idleTimeout", Integer.toString(5000))),
             VerifyHelper.POSITIVE, "launcher.postgresql.idleTimeout can't be <= 5000");
