+++ LaunchServer/src/main/java/pro/gravit/launchserver/auth/SQLSourceConfig.java
@@ -1,0 +1,9 @@
+package pro.gravit.launchserver.auth;
+
+import java.sql.Connection;
+import java.sql.SQLException;
+
+public interface SQLSourceConfig {
+    Connection getConnection() throws SQLException;
+    void close();
+}
